move casing utilities to its own module
